wrote initial thoughts and separated the code
